Eliminated compilation warning.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@22750 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/video/render3d/shader/shaderplugins/softshader/scan_z.h
@@ -81,7 +81,7 @@ namespace cspluginSoftshader
     ZBufMode_ZEqual (const InterpolateScanlinePerspCommon& ipol, 
       uint32* zBuff) : ipol(ipol), zBuff(zBuff) {}
 
-    bool Test () { return ipol.Iz.GetFixed() == *zBuff; }
+    bool Test () { return (uint32)ipol.Iz.GetFixed() == *zBuff; }
     void Update () { }
     void Advance () { zBuff++; }
   };
